moved audition types into types folder
