+++ moby_app/src/types/scripts.ts
